Moved blocks out of `articles` and into `core` so that they're not just under any one "app"
