+++ etna/articles/models.py
@@ -28,6 +28,10 @@ from taggit.models import ItemBase, TagBase
 
 from etna.authors.models import AuthorPageMixin
 from etna.collections.models import TopicalPageMixin
+from etna.core.blocks import (
+    AuthorPromotedPagesBlock,
+    FeaturedCollectionBlock,
+)
 from etna.core.models import (
     BasePageWithIntro,
     ContentWarningMixin,
@@ -47,8 +51,6 @@ from etna.records.fields import RecordField
 
 from .blocks import (
     ArticlePageStreamBlock,
-    AuthorPromotedPagesBlock,
-    FeaturedCollectionBlock,
 )
