+++ etna/core/blocks/__init__.py
@@ -1,27 +1,32 @@
 from .base import SectionDepthAwareStructBlock
 from .cta import LargeCardLinksBlock
-from .featured_content import FeaturedRecordArticleBlock
+from .featured_content import FeaturedRecordArticleBlock, FeaturedCollectionBlock, RelatedItemBlock
 from .image import ContentImageBlock, ImageBlock, NoCaptionImageBlock
 from .page_chooser import APIPageChooserBlock
 from .page_list import PageListBlock
 from .paragraph import ParagraphBlock, ParagraphWithHeading
-from .promoted_links import AuthorPromotedLinkBlock, PromotedLinkBlock
+from .promoted_links import AuthorPromotedLinkBlock, PromotedLinkBlock, PromotedListBlock, PromotedItemBlock, AuthorPromotedPagesBlock
 from .quote import QuoteBlock
-from .section import SectionBlock
+from .section import SectionBlock, SubHeadingBlock
 
 __all__ = [
     "APIPageChooserBlock",
+    "AuthorPromotedPagesBlock",
     "ContentImageBlock",
     "FeaturedRecordArticleBlock",
+    "FeaturedCollectionBlock",
     "ImageBlock",
     "NoCaptionImageBlock",
     "PageListBlock",
     "ParagraphBlock",
     "ParagraphWithHeading",
     "PromotedLinkBlock",
+    "PromotedListBlock",
+    "PromotedItemBlock",
     "AuthorPromotedLinkBlock",
     "LargeCardLinksBlock",
     "QuoteBlock",
+    "RelatedItemBlock",
     "SectionBlock",
     "SectionDepthAwareStructBlock",
     "SubHeadingBlock",
