+++ etna/core/blocks/section.py
@@ -2,6 +2,8 @@ from django.utils.text import slugify
 
 from wagtail import blocks
 
+from .base import SectionDepthAwareStructBlock
+
 
 class SectionBlock(blocks.StructBlock):
     heading = blocks.CharBlock(
@@ -20,3 +22,12 @@ class SectionBlock(blocks.StructBlock):
         context = super().get_context(value, parent_context)
         context["heading_id"] = self.get_heading_id(value)
         return context
+
+
+class SubHeadingBlock(SectionDepthAwareStructBlock):
+    heading = blocks.CharBlock(max_length=100, label="Sub-heading")
+
+    class Meta:
+        icon = "heading"
+        label = "Sub-heading"
+        template = "articles/blocks/sub_heading.html"
